Reorganization of the project
