+++ src/fr/dornacraft/servertools/commands/info/CmdGc.java
@@ -13,8 +13,8 @@ import org.bukkit.entity.Player;
 import org.bukkit.plugin.java.JavaPlugin;
 
 import fr.dornacraft.servertools.ServerTools;
-import fr.dornacraft.servertools.ServerToolsConfig;
 import fr.dornacraft.servertools.utils.Lag;
+import fr.dornacraft.servertools.utils.ServerToolsConfig;
 import fr.voltariuss.simpledevapi.MessageLevel;
 import fr.voltariuss.simpledevapi.UtilsAPI;
 import fr.voltariuss.simpledevapi.cmds.DornacraftCommand;
